+++ src/main/java/site/moamoa/backend/converter/postimage/PostImageConverter.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.converter;
+package site.moamoa.backend.converter.postimage;
 
 import java.util.List;
 import java.util.stream.Collectors;
