+++ src/main/java/site/moamoa/backend/exception/member/MemberIsNotAuthorException.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.exception;
+package site.moamoa.backend.exception.member;
 
 import site.moamoa.backend.api_payload.code.status.ErrorStatus;
 import site.moamoa.backend.global.exception.GeneralException;
