+++ src/main/java/site/moamoa/backend/global/oauth2/service/CustomOAuth2UserService.java
@@ -13,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import site.moamoa.backend.domain.Member;
 import site.moamoa.backend.global.oauth2.CustomOAuth2User;
 import site.moamoa.backend.global.oauth2.OAuthAttributes;
-import site.moamoa.backend.repository.MemberRepository;
+import site.moamoa.backend.repository.member.MemberRepository;
 import site.moamoa.backend.web.dto.request.MemberRequestDTO;
 
 import java.util.Collections;
