+++ src/main/java/site/moamoa/backend/repository/memberpost/MemberPostRepository.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.repository;
+package site.moamoa.backend.repository.memberpost;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
