+++ src/main/java/site/moamoa/backend/repository/post/PostRepository.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.repository;
+package site.moamoa.backend.repository.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
