+++ src/main/java/site/moamoa/backend/service/memberpost/MemberPostCommandService.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.service;
+package site.moamoa.backend.service.memberpost;
 
 import site.moamoa.backend.domain.mapping.MemberPost;
