+++ src/main/java/site/moamoa/backend/web/controller/PostController.java
@@ -18,7 +18,7 @@ import site.moamoa.backend.web.dto.base.AuthInfoDTO;
 import site.moamoa.backend.web.dto.request.PostRequestDTO.AddPost;
 import site.moamoa.backend.web.dto.request.PostRequestDTO.UpdatePostInfo;
 import site.moamoa.backend.web.dto.response.PostResponseDTO.*;
-import site.moamoa.backend.service.PostCommandService;
+import site.moamoa.backend.service.post.PostCommandService;
 
 @Tag(name = "공동구매 게시글 API", description = "공동구매 페이지 관련 API")
 @RequiredArgsConstructor
